requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
pypdf2>=3.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
matplotlib>=3.7.0
edge-tts>=6.1.0
azure-cognitiveservices-speech>=1.34.0
python-dotenv>=1.0.0
numpy>=1.24.0
pillow>=10.0.0
